+++ export/amsdos/amsdos.go
@@ -33,7 +33,7 @@ func SaveAmsdosFile(filename, extension string, data []byte, fileType, user byte
 	}
 	cpcFilename := AmsdosFilename(filename, extension)
 	copy(header.Filename[:], strings.Replace(cpcFilename, ".", "", -1))
-	header.Checksum = uint16(header.ComputedChecksum16())
+	header.Checksum = header.ComputedChecksum16()
 	log.GetLogger().Error("filesize:%d,#%.2x\n", filesize, filesize)
 	log.GetLogger().Error("Data length %d\n", binary.Size(data))
 	fw, err := os.Create(filename)
